- Refactored packages for exceptions and handlers - Renamed NotFoundException to EntityNotFoundException
