--- src/main/java/com/develhope/Java27_progetto3_team2/exception/NotFoundException.java
@@ -1,8 +1,0 @@
-package com.develhope.Java27_progetto3_team2.exception;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
-
