+++ src/main/java/com/develhope/Java27_progetto3_team2/exception/exceptions/CourierAlreadyExistException.java
@@ -1,4 +1,4 @@
-package com.develhope.Java27_progetto3_team2.exception;
+package com.develhope.Java27_progetto3_team2.exception.exceptions;
 
 public class CourierAlreadyExistException extends RuntimeException{
     public CourierAlreadyExistException(String message){
