+++ src/main/java/com/develhope/Java27_progetto3_team2/exception/exceptions/EntityNotFoundException.java
@@ -1,0 +1,8 @@
+package com.develhope.Java27_progetto3_team2.exception.exceptions;
+
+public class EntityNotFoundException extends RuntimeException {
+    public EntityNotFoundException(String message) {
+        super(message);
+    }
+}
+
