+++ src/main/java/com/develhope/Java27_progetto3_team2/exception/handlers/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
-package com.develhope.Java27_progetto3_team2.exception;
+package com.develhope.Java27_progetto3_team2.exception.handlers;
 
+import com.develhope.Java27_progetto3_team2.exception.exceptions.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -9,8 +10,8 @@ import org.springframework.web.context.request.WebRequest;
 @ControllerAdvice
 public class GlobalExceptionHandler {
 
-    @ExceptionHandler(NotFoundException.class)
-    public ResponseEntity<String> handleNotFoundException(NotFoundException ex, WebRequest request) {
+    @ExceptionHandler(EntityNotFoundException.class)
+    public ResponseEntity<String> handleNotFoundException(EntityNotFoundException ex, WebRequest request) {
         return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
     }
